Barebones UI changes in Code Editor
