+++ src/CodeEditor.css
@@ -1,37 +1,85 @@
 /* CodeEditor.css */
-
-.editor-container {
+body{
+  background-color: #1e1e1e;
+  color: #f5f5f5;
+}
+nav{
+  background-color: #1e1e1e;
+  color: white;
+  margin: 0%;
+  height: 8vh;
+  padding: 1rem;
   display: flex;
-  margin: 20px;
+  align-items: center;
+}
+nav h2{
+  margin: 0;
+}
+
+.canvas{
+  display: flex;
+  width: 100vw;
+}
+
+.question-container{
+  width: 35vw;
+  margin: 1rem;
+  height: 70vh;
+  padding: 0 1vw;
+  overflow: scroll;
   border: 1px solid #ccc;
   border-radius: 5px;
-  height:60vh
+}
+.question-container h3{
+  font-size: 2rem;
+}
+.question-container p{
+  line-height: 2rem;
+}
+
+.editor-container {
+  margin: 1rem 1rem 1rem 0;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  width:70vw;
+  height:70vh;
 }
 
 .editor-sidebar {
   display: flex;
-  flex-direction: column;
   padding: 10px;
-  background-color: #f5f5f5;
-  border-right: 1px solid #ccc;
+  background-color: #1e1e1e;
+  border-bottom: 1px solid #ccc;
 }
 
 .editor-sidebar select {
   margin-bottom: 10px;
-  padding: 5px;
+  background-color: #1e1e1e;
+  color: #f5f5f5;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  width: 20rem;
+  padding:0.5rem;
+  height: 2.5rem;
 }
 
 .editor-sidebar button {
-  padding: 10px;
-  background-color: #0078d4;
+  padding: 0.5rem;
+  height: 2.5rem;
+  width: 5rem;
+  margin-left: 1rem;
+  background-color: #1e1e1e;
   color: #fff;
   border: none;
+  border: 1px solid #ccc;
   border-radius: 5px;
   cursor: pointer;
 }
 
 .editor-content {
-  width:100%
+  padding:1rem;
+  height: 55vh;
+  overflow-y: hidden;
 }
 
 .output-container {
